vacation facade
similar to the user facade, i wrote some functions that get and check the users input, and pass the params to the logic. for eample, i wrote a function called delete_vacation which gets the id of a vacation, checks if this id matches any id in the db and sends the id to the function in the vacationlogic that delets the vacation.